Fjern gammel behandlingsresultat toggle og gammel kode
