+++ src/main/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/behandlingsresultat/BehandlingsresultatValideringUtils.kt
@@ -6,7 +6,12 @@ import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingType
 import no.nav.familie.ks.sak.kjerne.behandling.domene.Behandlingsresultat
 import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.PersonResultat
+import no.nav.familie.ks.sak.kjerne.beregning.EndretUtbetalingAndelMedAndelerTilkjentYtelse
+import no.nav.familie.ks.sak.kjerne.beregning.TilkjentYtelseValidator
+import no.nav.familie.ks.sak.kjerne.beregning.domene.TilkjentYtelse
+import no.nav.familie.ks.sak.kjerne.endretutbetaling.EndretUtbetalingAndelValidator
 import no.nav.familie.ks.sak.kjerne.personident.Aktør
+import no.nav.familie.ks.sak.kjerne.personopplysninggrunnlag.domene.PersonopplysningGrunnlag
 
 object BehandlingsresultatValideringUtils {
     internal fun validerAtBarePersonerFremstiltKravForEllerSøkerHarFåttEksplisittAvslag(
@@ -21,6 +26,19 @@ object BehandlingsresultatValideringUtils {
                 melding = "Det eksisterer personer som har fått eksplisitt avslag, men som det ikke har blitt fremstilt krav for.",
             )
         }
+    }
+
+    fun validerAtBehandlingsresultatKanUtføres(
+        personopplysningGrunnlag: PersonopplysningGrunnlag,
+        tilkjentYtelse: TilkjentYtelse,
+        endretUtbetalingMedAndeler: List<EndretUtbetalingAndelMedAndelerTilkjentYtelse>,
+    ) {
+        // valider TilkjentYtelse
+        TilkjentYtelseValidator.validerAtTilkjentYtelseHarFornuftigePerioderOgBeløp(tilkjentYtelse, personopplysningGrunnlag)
+
+        // valider EndretUtbetalingAndel
+        EndretUtbetalingAndelValidator.validerAtAlleOpprettedeEndringerErUtfylt(endretUtbetalingMedAndeler.map { it.endretUtbetaling })
+        EndretUtbetalingAndelValidator.validerAtEndringerErTilknyttetAndelTilkjentYtelse(endretUtbetalingMedAndeler)
     }
 
     internal fun validerBehandlingsresultat(
@@ -58,6 +76,27 @@ object BehandlingsresultatValideringUtils {
                 "Behandlingsårsak ${behandling.opprettetÅrsak.visningsnavn.lowercase()} " +
                     "er ugyldig i kombinasjon med resultat '${resultat.displayName.lowercase()}'."
             throw FunksjonellFeil(frontendFeilmelding = feilmelding, melding = feilmelding)
+        }
+    }
+
+    fun validerUtledetBehandlingsresultat(
+        behandling: Behandling,
+        behandlingsresultat: Behandlingsresultat,
+    ) {
+        when {
+            behandling.type !in behandlingsresultat.gyldigeBehandlingstyper -> {
+                val feilmelding =
+                    "Behandlingsresultatet ${behandlingsresultat.displayName.lowercase()} " +
+                        "er ugyldig i kombinasjon med behandlingstype '${behandling.type.visningsnavn}'."
+                throw FunksjonellFeil(frontendFeilmelding = feilmelding, melding = feilmelding)
+            }
+
+            behandlingsresultat.erAvslått() && behandling.erKlage() -> {
+                val feilmelding =
+                    "Behandlingsårsak ${behandling.opprettetÅrsak.visningsnavn.lowercase()} " +
+                        "er ugyldig i kombinasjon med resultat '${behandlingsresultat.displayName.lowercase()}'."
+                throw FunksjonellFeil(frontendFeilmelding = feilmelding, melding = feilmelding)
+            }
         }
     }
 }
