+++ src/test/enhetstester/kotlin/no/nav/familie/ks/sak/kjerne/behandling/steg/behandlingsresultat/BehandlingsresultatServiceTest.kt
@@ -11,21 +11,15 @@ import no.nav.familie.ks.sak.api.dto.SøknadDto
 import no.nav.familie.ks.sak.common.util.LocalDateProvider
 import no.nav.familie.ks.sak.data.lagBehandling
 import no.nav.familie.ks.sak.data.lagPerson
-import no.nav.familie.ks.sak.data.lagPersonopplysningGrunnlag
-import no.nav.familie.ks.sak.data.lagVilkårsvurderingOppfylt
 import no.nav.familie.ks.sak.data.randomAktør
 import no.nav.familie.ks.sak.data.randomFnr
 import no.nav.familie.ks.sak.kjerne.behandling.BehandlingService
-import no.nav.familie.ks.sak.kjerne.behandling.domene.Behandling
 import no.nav.familie.ks.sak.kjerne.behandling.domene.BehandlingÅrsak
 import no.nav.familie.ks.sak.kjerne.behandling.steg.registrersøknad.SøknadGrunnlagService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.VilkårsvurderingService
 import no.nav.familie.ks.sak.kjerne.behandling.steg.vilkårsvurdering.domene.Vilkårsvurdering
-import no.nav.familie.ks.sak.kjerne.beregning.AndelerTilkjentYtelseOgEndreteUtbetalingerService
-import no.nav.familie.ks.sak.kjerne.beregning.EndretUtbetalingAndelMedAndelerTilkjentYtelse
 import no.nav.familie.ks.sak.kjerne.beregning.domene.AndelTilkjentYtelseRepository
 import no.nav.familie.ks.sak.kjerne.endretutbetaling.EndretUtbetalingAndelService
-import no.nav.familie.ks.sak.kjerne.endretutbetaling.domene.EndretUtbetalingAndel
 import no.nav.familie.ks.sak.kjerne.eøs.kompetanse.KompetanseService
 import no.nav.familie.ks.sak.kjerne.personident.Aktør
 import no.nav.familie.ks.sak.kjerne.personident.PersonidentService
@@ -42,9 +36,6 @@ class BehandlingsresultatServiceTest {
 
     @MockK
     private lateinit var behandlingService: BehandlingService
-
-    @MockK
-    private lateinit var andelerTilkjentYtelseOgEndreteUtbetalingerService: AndelerTilkjentYtelseOgEndreteUtbetalingerService
 
     @MockK
     private lateinit var vilkårsvurderingService: VilkårsvurderingService
@@ -69,121 +60,6 @@ class BehandlingsresultatServiceTest {
 
     @InjectMockKs
     private lateinit var behandlingsresultatService: BehandlingsresultatService
-
-    @Test
-    fun `lagBehandlingsresulatPersoner - skal returnere BehandlingsresultatPerson som ikke er eksplisittAvslag hvis det ikke er avslag i vilkårsvurdering eller endretutbetaling`() {
-        val behandling = mockk<Behandling>(relaxed = true)
-        val personopplysningGrunnlag =
-            lagPersonopplysningGrunnlag(behandlingId = behandling.id, barnasIdenter = listOf(randomFnr(), randomFnr()))
-        val vilkårsvurdering = lagVilkårsvurderingOppfylt(personopplysningGrunnlag.personer)
-
-        val andelerMedEndringer =
-            personopplysningGrunnlag.personer.map {
-                EndretUtbetalingAndelMedAndelerTilkjentYtelse(
-                    EndretUtbetalingAndel(
-                        behandlingId = behandling.id,
-                        erEksplisittAvslagPåSøknad = false,
-                        person = it,
-                    ),
-                    emptyList(),
-                )
-            }
-
-        every { personopplysningGrunnlagService.hentAktivPersonopplysningGrunnlagThrows(behandling.id) } returns personopplysningGrunnlag
-        every {
-            andelerTilkjentYtelseOgEndreteUtbetalingerService.finnEndreteUtbetalingerMedAndelerTilkjentYtelse(
-                behandlingId = behandling.id,
-            )
-        } returns andelerMedEndringer
-
-        val behandlingsresulatPersoner =
-            behandlingsresultatService.lagBehandlingsresulatPersoner(
-                behandling = behandling,
-                personerFremslitKravFor = emptyList(),
-                forrigeAndelerMedEndringer = emptyList(),
-                vilkårsvurdering = vilkårsvurdering,
-                andelerMedEndringer = emptyList(),
-            )
-
-        assertThat(behandlingsresulatPersoner.all { it.eksplisittAvslag }, Is(false))
-    }
-
-    @Test
-    fun `lagBehandlingsresulatPersoner - skal returnere BehandlingsresultatPerson som er eksplisittAvslag hvis det er avslag i endretutbetaling`() {
-        val behandling = mockk<Behandling>(relaxed = true)
-        val personopplysningGrunnlag =
-            lagPersonopplysningGrunnlag(behandlingId = behandling.id, barnasIdenter = listOf(randomFnr(), randomFnr()))
-        val vilkårsvurdering = lagVilkårsvurderingOppfylt(personopplysningGrunnlag.personer)
-
-        val andelerMedEndringer =
-            personopplysningGrunnlag.personer.map {
-                EndretUtbetalingAndelMedAndelerTilkjentYtelse(
-                    EndretUtbetalingAndel(
-                        behandlingId = behandling.id,
-                        erEksplisittAvslagPåSøknad = true,
-                        person = it,
-                    ),
-                    emptyList(),
-                )
-            }
-
-        every { personopplysningGrunnlagService.hentAktivPersonopplysningGrunnlagThrows(behandling.id) } returns personopplysningGrunnlag
-        every {
-            andelerTilkjentYtelseOgEndreteUtbetalingerService.finnEndreteUtbetalingerMedAndelerTilkjentYtelse(
-                behandlingId = behandling.id,
-            )
-        } returns andelerMedEndringer
-
-        val behandlingsresulatPersoner =
-            behandlingsresultatService.lagBehandlingsresulatPersoner(
-                behandling = behandling,
-                personerFremslitKravFor = emptyList(),
-                forrigeAndelerMedEndringer = emptyList(),
-                vilkårsvurdering = vilkårsvurdering,
-                andelerMedEndringer = emptyList(),
-            )
-
-        assertThat(behandlingsresulatPersoner.all { it.eksplisittAvslag }, Is(true))
-    }
-
-    @Test
-    fun `lagBehandlingsresulatPersoner - skal returnere BehandlingsresultatPerson som er eksplisittAvslag hvis det er avslag i vilkårsvurderingen`() {
-        val behandling = mockk<Behandling>(relaxed = true)
-        val personopplysningGrunnlag =
-            lagPersonopplysningGrunnlag(behandlingId = behandling.id, barnasIdenter = listOf(randomFnr(), randomFnr()))
-        val vilkårsvurdering =
-            lagVilkårsvurderingOppfylt(personer = personopplysningGrunnlag.personer, erEksplisittAvslagPåSøknad = true)
-
-        val andelerMedEndringer =
-            personopplysningGrunnlag.personer.map {
-                EndretUtbetalingAndelMedAndelerTilkjentYtelse(
-                    EndretUtbetalingAndel(
-                        behandlingId = behandling.id,
-                        erEksplisittAvslagPåSøknad = false,
-                        person = it,
-                    ),
-                    emptyList(),
-                )
-            }
-
-        every { personopplysningGrunnlagService.hentAktivPersonopplysningGrunnlagThrows(behandling.id) } returns personopplysningGrunnlag
-        every {
-            andelerTilkjentYtelseOgEndreteUtbetalingerService.finnEndreteUtbetalingerMedAndelerTilkjentYtelse(
-                behandlingId = behandling.id,
-            )
-        } returns andelerMedEndringer
-
-        val behandlingsresulatPersoner =
-            behandlingsresultatService.lagBehandlingsresulatPersoner(
-                behandling = behandling,
-                personerFremslitKravFor = emptyList(),
-                forrigeAndelerMedEndringer = emptyList(),
-                vilkårsvurdering = vilkårsvurdering,
-                andelerMedEndringer = emptyList(),
-            )
-
-        assertThat(behandlingsresulatPersoner.all { it.eksplisittAvslag }, Is(true))
-    }
 
     @Test
     fun `finnPersonerFremstiltKravFor skal returnere tom liste dersom behandlingen ikke er søknad, fødselshendelse eller manuell migrering`() {
